chore: convert user service to hexagonal architecture
