+++ user-service/src/main/java/com/financialmonitoring/userservice/adapter/security/constants/SecurityConstants.java
@@ -1,0 +1,11 @@
+package com.financialmonitoring.userservice.adapter.security.constants;
+
+public class SecurityConstants {
+
+    public static final String[] PUBLIC_ROUTES =
+            new String[] {
+                "/api/v1/auth/register", "/api/v1/auth/login", "/api/v1/auth/verify-token"
+            };
+
+    public static final int BCRYPT_SALT = 10;
+}
