--- user-service/src/main/java/com/financialmonitoring/userservice/core/controller/AuthController.java
@@ -1,44 +1,0 @@
-package com.financialmonitoring.userservice.core.controller;
-
-import com.financialmonitoring.userservice.core.dto.LoginRequestDTO;
-import com.financialmonitoring.userservice.core.dto.LoginResponseDTO;
-import com.financialmonitoring.userservice.core.dto.RegisterRequestDTO;
-import com.financialmonitoring.userservice.core.dto.RegisterResponseDTO;
-import com.financialmonitoring.userservice.core.service.UserService;
-import jakarta.validation.Valid;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/v1/auth")
-public class AuthController {
-
-    private final UserService userService;
-
-    public AuthController(UserService userService) {
-        this.userService = userService;
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<LoginResponseDTO> login(@Valid @RequestBody LoginRequestDTO requestDTO) {
-        return ResponseEntity.ok(userService.login(requestDTO));
-    }
-
-    @PostMapping("/register")
-    public ResponseEntity<RegisterResponseDTO> register(@Valid @RequestBody RegisterRequestDTO requestDTO) {
-        return ResponseEntity.ok(userService.register(requestDTO));
-    }
-
-    @GetMapping("/verify-token")
-    public ResponseEntity<?> verifyToken(@RequestHeader("Authorization") String authHeader) {
-        return userService.verifyToken(authHeader)
-                ? ResponseEntity.ok().build()
-                : ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-    }
-}
