--- user-service/src/main/java/com/financialmonitoring/userservice/core/dto/RegisterResponseDTO.java
@@ -1,40 +1,0 @@
-package com.financialmonitoring.userservice.core.dto;
-
-public class RegisterResponseDTO {
-    private Long id;
-    private String firstName;
-    private String email;
-
-    public RegisterResponseDTO() {
-    }
-
-    public RegisterResponseDTO(Long id, String firstName, String email) {
-        this.id = id;
-        this.firstName = firstName;
-        this.email = email;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
