--- user-service/src/main/java/com/financialmonitoring/userservice/core/model/Role.java
@@ -1,84 +1,0 @@
-package com.financialmonitoring.userservice.core.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.SequenceGenerator;
-import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
-import java.io.Serial;
-import java.io.Serializable;
-import java.util.Objects;
-import org.springframework.security.core.GrantedAuthority;
-
-@Entity
-@Table(
-        name = "tb_roles",
-        uniqueConstraints = {
-            @UniqueConstraint(
-                    name = "unique_role_name",
-                    columnNames = {"name"})
-        })
-@SequenceGenerator(name = "seq_role", sequenceName = "seq_role", allocationSize = 1)
-public class Role implements GrantedAuthority, Serializable {
-
-    @Serial private static final long serialVersionUID = 4403061068892226017L;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_role")
-    private Long id;
-
-    @Column(nullable = false, unique = true, length = 50)
-    private String name;
-
-    public Role() {}
-
-    public Role(String name) {
-        this.name = name;
-    }
-
-    public Role(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (o == null || getClass() != o.getClass()) return false;
-        Role role = (Role) o;
-        return Objects.equals(id, role.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(id);
-    }
-
-    @Override
-    public String toString() {
-        return "Role{" + "name='" + name + '\'' + '}';
-    }
-
-    @Override
-    public String getAuthority() {
-        return getName();
-    }
-}
