--- user-service/src/main/java/com/financialmonitoring/userservice/core/model/User.java
@@ -1,288 +1,0 @@
-package com.financialmonitoring.userservice.core.model;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.ConstraintMode;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.ForeignKey;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.JoinTable;
-import jakarta.persistence.ManyToMany;
-import jakarta.persistence.SequenceGenerator;
-import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
-import java.io.Serial;
-import java.io.Serializable;
-import java.time.LocalDateTime;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Objects;
-import java.util.Set;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-@Entity
-@Table(
-        name = "tb_users",
-        uniqueConstraints = {@UniqueConstraint(name = "unique_email", columnNames = "email")})
-@SequenceGenerator(name = "seq_user", sequenceName = "seq_user", allocationSize = 1)
-public class User implements UserDetails, Serializable {
-    @Serial private static final long serialVersionUID = -2288539263513153294L;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_user")
-    @Column(nullable = false)
-    private Long id;
-
-    @Column(nullable = false, length = 100)
-    private String firstName;
-
-    private String lastName;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @ManyToMany(fetch = FetchType.EAGER, targetEntity = Role.class, cascade = CascadeType.DETACH)
-    @JoinTable(
-            name = "tb_users_roles",
-            uniqueConstraints = {
-                @UniqueConstraint(
-                        columnNames = {"user_id", "role_id"},
-                        name = "unique_user_role"),
-            },
-            joinColumns = {
-                @JoinColumn(
-                        name = "user_id",
-                        referencedColumnName = "id",
-                        table = "tb_users",
-                        foreignKey =
-                                @ForeignKey(name = "user_fk", value = ConstraintMode.CONSTRAINT))
-            },
-            inverseJoinColumns = {
-                @JoinColumn(
-                        name = "role_id",
-                        referencedColumnName = "id",
-                        table = "tb_roles",
-                        foreignKey =
-                                @ForeignKey(name = "role_fk", value = ConstraintMode.CONSTRAINT))
-            })
-    private Set<Role> roles = new HashSet<>();
-
-    @Column(nullable = false)
-    @CreationTimestamp
-    private LocalDateTime createdAt;
-
-    @Column(updatable = false)
-    @UpdateTimestamp
-    private LocalDateTime updatedAt;
-
-    @Column(nullable = false)
-    private Boolean isActive;
-
-    public User() {}
-
-    public User(Builder builder) {
-        this.id = builder.id;
-        this.firstName = builder.firstName;
-        this.lastName = builder.lastName;
-        this.email = builder.email;
-        this.password = builder.password;
-        this.roles = builder.roles;
-        this.isActive = builder.isActive;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return this.roles;
-    }
-
-    @Override
-    public String getPassword() {
-        return this.password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    @Override
-    public String getUsername() {
-        return this.email;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public Set<Role> getRoles() {
-        return roles;
-    }
-
-    public void setRoles(Set<Role> roles) {
-        this.roles = roles;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public LocalDateTime getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(LocalDateTime updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-
-    public Boolean getActive() {
-        return isActive;
-    }
-
-    public void setActive(Boolean active) {
-        isActive = active;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (o == null || getClass() != o.getClass()) return false;
-        User user = (User) o;
-        return Objects.equals(id, user.id) && Objects.equals(email, user.email);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, email);
-    }
-
-    @Override
-    public String toString() {
-        return "User{"
-                + "id="
-                + id
-                + ", firstName='"
-                + firstName
-                + '\''
-                + ", lastName='"
-                + lastName
-                + '\''
-                + ", email='"
-                + email
-                + '\''
-                + '}';
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    public static class Builder {
-
-        private Long id;
-        private String firstName;
-        private String lastName;
-        private String email;
-        private String password;
-        private Set<Role> roles;
-        public Boolean isActive;
-
-        public Builder id(Long id) {
-            this.id = id;
-            return this;
-        }
-
-        public Builder firstName(String firstName) {
-            this.firstName = firstName;
-            return this;
-        }
-
-        public Builder lastName(String lastName) {
-            this.lastName = lastName;
-            return this;
-        }
-
-        public Builder email(String email) {
-            this.email = email;
-            return this;
-        }
-
-        public Builder password(String password) {
-            this.password = password;
-            return this;
-        }
-
-        public Builder roles(Set<Role> roles) {
-            this.roles = roles;
-            return this;
-        }
-
-        public Builder active(Boolean isActive) {
-            this.isActive = isActive;
-            return this;
-        }
-
-        public User build() {
-            return new User(this);
-        }
-    }
-}
