--- user-service/src/main/java/com/financialmonitoring/userservice/core/repository/RoleRepository.java
@@ -1,16 +1,0 @@
-package com.financialmonitoring.userservice.core.repository;
-
-import com.financialmonitoring.userservice.core.model.Role;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-
-    @Query(
-            value = "SELECT * FROM tb_roles r WHERE UPPER(r.name) = UPPER(:name)",
-            nativeQuery = true)
-    Optional<Role> findByName(String name);
-}
