+++ user-service/src/main/java/com/financialmonitoring/userservice/domain/port/in/GetUserByEmailUseCase.java
@@ -1,0 +1,7 @@
+package com.financialmonitoring.userservice.domain.port.in;
+
+import com.financialmonitoring.userservice.adapter.out.entity.User;
+
+public interface GetUserByEmailUseCase {
+    User getUserByEmail(String email);
+}
