+++ user-service/src/main/java/com/financialmonitoring/userservice/domain/port/in/LoginUseCase.java
@@ -1,0 +1,8 @@
+package com.financialmonitoring.userservice.domain.port.in;
+
+import com.financialmonitoring.userservice.domain.dto.LoginRequestDTO;
+import com.financialmonitoring.userservice.domain.dto.LoginResponseDTO;
+
+public interface LoginUseCase {
+    LoginResponseDTO login(LoginRequestDTO requestDTO);
+}
