+++ user-service/src/main/java/com/financialmonitoring/userservice/domain/port/in/RegisterUserUseCase.java
@@ -1,0 +1,8 @@
+package com.financialmonitoring.userservice.domain.port.in;
+
+import com.financialmonitoring.userservice.domain.dto.RegisterRequestDTO;
+import com.financialmonitoring.userservice.domain.dto.RegisterResponseDTO;
+
+public interface RegisterUserUseCase {
+    RegisterResponseDTO register(RegisterRequestDTO requestDTO);
+}
