+++ user-service/src/main/java/com/financialmonitoring/userservice/domain/port/in/VerifyTokenUseCase.java
@@ -1,0 +1,5 @@
+package com.financialmonitoring.userservice.domain.port.in;
+
+public interface VerifyTokenUseCase {
+    boolean verifyToken(String authHeader);
+}
